refactor BDPT to multiple files
